fix(types): make lume-js/state type-check without lib.dom
Review finding: src/state.d.ts re-exported from index.d.ts, whose bindDom/Handler/template declarations reference HTMLElement - so a Node/worker consumer with lib:["ES2020"] and skipLibCheck:false got "Cannot find name 'HTMLElement'" from the universal entry's own type graph, the exact audience the entry targets (runtime was always clean; the repo tsconfig masked it via lib DOM + skipLibCheck).

Inverted to mirror the runtime module graph: state.d.ts is now the self-contained kernel source of truth (Unsubscribe, Subscriber, Plugin, TypedPlugin, ReactiveState, state, batch, withReadObserver) and index.d.ts re-exports it, adding only the DOM-flavored API (bindDom, Handler, effect). Verified: state.d.ts compiles under lib:["ES2020"] with skipLibCheck off and zero ambient types; the full repo typecheck stays green; addons/handlers d.ts imports resolve unchanged through the re-exports.
